+++ modules/app-core/src/commonMain/kotlin/org/archivekeep/app/core/persistence/drivers/filesystem/operations/AddFileSystemRepositoryUseCase.kt
@@ -1,0 +1,12 @@
+package org.archivekeep.app.core.persistence.drivers.filesystem.operations
+
+import kotlinx.coroutines.CoroutineScope
+import org.archivekeep.app.core.persistence.drivers.filesystem.FileSystemStorageType
+
+interface AddFileSystemRepositoryUseCase {
+    fun begin(
+        scope: CoroutineScope,
+        path: String,
+        intendedStorageType: FileSystemStorageType?,
+    ): AddFileSystemRepositoryOperation
+}