+++ modules/app-ui/src/androidMain/kotlin/org/archivekeep/app/ui/dialogs/repository/registry/AddFileSystemRepositoryDialog.android.kt
@@ -12,45 +12,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LifecycleEventEffect
-import io.github.vinceglb.filekit.compose.rememberDirectoryPickerLauncher
-import io.github.vinceglb.filekit.core.FileKitPlatformSettings
-import kotlinx.coroutines.runBlocking
 import org.archivekeep.app.core.persistence.drivers.filesystem.MountedFileSystem
-import org.archivekeep.app.ui.domain.wiring.LocalFileStores
 import java.io.File
-
-@Composable
-actual fun filesystemRepositoryDirectoryPicker(onResult: (result: PickResult) -> Unit): () -> Unit {
-    val fileStores = LocalFileStores.current
-
-    val pickerLauncher =
-        rememberDirectoryPickerLauncher(
-            title = "Pick a directory",
-            platformSettings = FileKitPlatformSettings(),
-        ) { directory ->
-            if (directory == null) {
-                onResult(PickResult.Failure(RuntimeException("Directory picker returned null directory")))
-            } else {
-                val path = directory.path
-
-                if (path == null) {
-                    onResult(PickResult.Failure(RuntimeException("Path not present for $directory")))
-                } else {
-                    onResult(
-                        PickResult.Success(
-                            getPathFromURI(
-                                // TODO - refactor to be executed in appropriate suspend execution context
-                                runBlocking { fileStores.loadFreshMountPoints() },
-                                directory.uri,
-                            ),
-                        ),
-                    )
-                }
-            }
-        }
-
-    return pickerLauncher::launch
-}
 
 @Composable
 actual fun platformSpecificFileSystemRepositoryGuard(): PlatformSpecificPermissionFulfilment = storageManagerGuard()